url: Have isUrlOnRealm take a URL object, not a string; simplify
This fixes a false negative; see the removed implementation comment
in the function.

Now this function uses the same check we use in the 5 or so other
places where we check if a URL is on a given realm. As a comment
says, we should probably either reuse this function in all those
cases, or delete it.