homescreen_tabs: Change "globe" icon to "align-left" icon.
This change was made to make the All-Messages icon in mobile Zulip congruent to the one found in browser Zulip.

This change was performed by changing the "name" property of the first TopTabButton component found in HomeScreen.js. This property was changed from "globe" to "align-left".

It is important to note that this property is a string restricted by Flow's literal union type. A list of all valid strings for this property can be found in "FeatherGlyphs" type at the top of Icons.js.

This commit has passed all included tests.

Fixes: #5303.