notifications: Remove Android summary notification.
Removes the summary notification from the notification group in
NotificationManager.kt, and removes logic around removing the
summary notification when Zulip messages are read.

This change allows the user to tap on the body of the
notification to open the app, which was not possible with the summary
notification (only the chevron would expand the
notification, which would then allow the user to interact with the
individual notifications below it.

Please note that removing the summary notification breaks proper
grouping: multiple private messages from the same user will correctly
be grouped together, but notifications for different stream
conversations appear separately.
This was tested manually.

Fixes part of #5242.