internalLinks [nfc]: Inline isMessageLink
The logic here is sloppy: just because a narrow link includes "near"
doesn't mean it uses the "near" operator, i.e., the operator that's
meant to point to (or near) a specific message. It could be a
non-"near" narrow link that just happens to point to a stream with
"near" in its slugified name. Or the realm's base URL could have
"near" in it. Think of discussions about li"near" algebra, for
example.

Better would be to inspect the result of `getPathsfromUrl`, which,
following this inlining, is now in scope as `paths`. So we'll do
that soon.

We could instead have left the function definition and had it take
`paths` as an argument… But I don't think the question "does this
link use the 'near' operator" is likely to be useful except where we
want to grab the corresponding operand, as we do in
getMessageIdFromLink.

Inline the tests for the old isMessageLink into the tests for
getMessageIdFromLink, with appropriate test descriptions.