compose: Allow image multi-selection, where supported
We recently completed the feature of not immediately sending
messages with image uploads, in #5474. With that, it became possible
to send a message with multiple uploaded images in it. But you had
to go through the image-selection flow once for each of your chosen
images; see discussion:
  https://github.com/zulip/zulip-mobile/pull/5474#pullrequestreview-1246356066

This gives a smoother experience by letting you choose multiple
images in one image-selection session.

Related: #2366
Co-authored-by: Akash Dhiman <akash.d0407@gmail.com>