homescreen_icon [nfc]: Swap TopTabButton with TopTabButtonGeneral.
to allow button to consume SpecificIconType, a component type that can
return a FontAwesome icon.

The current component returns icons from the Feather collection.
We want the new icon to be an exact match to the one found
in web Zulip, which is from the FontAwesome collection, but we
need the SpecificIconType component in order to do so.