notifications: Open to unreads from summary notification
Adds an intent to the summary notification. When the summary
notification is clicked in the collapsed state, the app will open and
reset itself to the unreads screen. It will also switch to the
appropriate account if needed.

Because the intent flags (Intent.FLAG_ACTIVITY_NEW_TASK
and Intent.FLAG_ACTIVITY_CLEAR_TOP) apply to both the normal
notification and the summary one (and are prefaced with explanatory
comments), I moved them to a variable before where both notifications
are created.

This was manually tested for the following cases, with both the app
already being on the correct account and being on a different account:
- App open in background on non-unreads screen
- App open in background on unreads screen
- App not open in background

I also made sure that tapping on a notification in the expanded state
still routes the user to the correct screen.

Fixes #5242.