ComposeBox types: Use proper type for `ref` instead of $FlowFixMe
The bug described in the comment doesn't seem to be active anymore.
If I say `const a = ref.current`, then ask `yarn flow type-at-pos`
what type it has inferred for `a`, it gives

  null
  |
  {
    blur(): void,
    +clear: () => void,
    focus(): void,
    +getNativeRef: () => ?ElementRef<HostComponent<mixed>>,
    +isFocused: () => boolean,
    measure(callback: MeasureOnSuccessCallback): void,
    measureInWindow(callback: MeasureInWindowOnSuccessCallback): void,
    measureLayout(
      relativeToNativeNode: (number | ElementRef<HostComponent<mixed>>),
      onSuccess: MeasureLayoutOnSuccessCallback,
      onFail?: () => void
    ): void,
    setNativeProps(nativeProps: {...}): void
  }

which is a lot better than `any`, and nothing in it stands out to me
as obviously wrong.