ComposeBox [nfc]: Pull out InputState type
Soon, we want to bring useUncontrolledInput out into a utils file,
both because the hook designed to be reusable and because doing so
will let us shrink the large ComposeBox file by several dozen lines.

When we put it in a utils file, we'll need to annotate its return
type, and for that, it'll be convenient to have this reusable type.

Also rename its param from `args` to `initialState`, since the
initial value and selection are the only inputs the function
currently takes, and it's better to think of them together as "the
state", as we do elsewhere, than separately.

Also with an aim to make this hook stand alone as a util, spell out
the InputSelection type literally. We don't mean to draw a
connection between this hook and other things that use that type,
like autocomplete UIs.