i18n [nfc]: Remove unused "Subscriptions" string from messages_en.json
`git grep` confirms that this string doesn't actually appear in UI
code. See the investigation that flagged this string:
  https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/Strings.20in.20messages_en.2Ejson.20but.20not.20in.20the.20app/near/1425314

The string looks to have been present in some version of this file
since commit ab19d1ba4. But it stopped appearing in a UI component
in 9fd916225.

…And the latter commit even says the UI component wasn't actually
getting used at the time.