i18n [nfc]: Remove unused "Search people" string from messages_en.json
`git grep` confirms that this string doesn't actually appear in UI
code. See the investigation that flagged this string:
  https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/Strings.20in.20messages_en.2Ejson.20but.20not.20in.20the.20app/near/1425314

The string looks to have been present in some version of this file
since commit ab19d1ba4. At that time, it was used for a button to
start a PM conversation.

In b7a246123, that single button was replaced with two buttons, one
for 1:1 PMs and one for group PMs. Those two buttons now say "New
PM" and "New group PM", and those strings are represented in
messages_en.json, as they should be.