safe areas: Put OfflineNotice on each tab screen, out of top inset
This fixes a regression introduced in fc06c9466. In that commit, we
asked MainTabsScreen to stop padding the top inset, and asked each
individual tab screen to pad it. With this <OfflineNotice /> being
left behind in MainTabsScreen, it lost its protection from the top
inset, and on my device the crucial "No Internet connection" text
was completely hidden by the notch.

So, remove that single <OfflineNotice /> and add one in each of the
four tab screens, using each screen's protection against rendering
content in the top inset.