offline notice [nfc]: Put <OfflineNotice />s into nearby *NavBar components
We want to try making the OfflineNotice a surface that fills the top
inset, instead of being a bar that cuts through the middle of an
existing surface.

With that goal in mind, move our `<OfflineNotice />`s closer to the
UI code responsible for handling the top inset.

But without making any functional changes yet: in ChatScreen, for
example, the `<OfflineNotice />` goes from being the next element
down from `<ChatNavBar />`, to being the last element *in*
ChatNavBar, in a way that ensures the visual layout doesn't change.

And likewise everywhere else an `<OfflineNotice />` was just below a
`<ModalNavBar />` or a `<ModalSearchNavBar />`.

We'll start changing the layout next.