android build: Upgrade Gradle to 7.0.2
Done to follow the template-app change in
facebook/react-native@13107fa3d. (Which it looks like RN actually
intended to make in facebook/react-native@85249cafe but forgot;
that's not important, I think.)

We didn't take facebook/react-native@13107fa3d on the path to the RN
v0.67 upgrade because of a blocker that was resolved with the Expo
44 upgrade in expo/expo@3edc37ae4. For details, expand the
`expo-screen-orientation` block in
  https://github.com/zulip/zulip-mobile/pull/5203#issuecomment-1020722398
and see the comment on the `Installing ExpoModulesCore 0.6.4` output
from `pod install`.

Done by following the "How to upgrade Gradle" instructions at the
top of gradle-wrapper.properties. The result matched the
template-app commit except for upstream's line endings got messed
with because they didn't have Git set up properly to keep them
consistent:
  https://github.com/zulip/zulip-mobile/pull/5426#issuecomment-1166078431 .