android build: Upgrade Gradle to 7.2
Done to follow the template-app change in
facebook/react-native@9ae336743. That commit was released with RN
v0.67, but we didn't take it on the path to that upgrade; see
previous commit for details.

Done by following the "How to upgrade Gradle" instructions at the
top of gradle-wrapper.properties. This made a bunch of changes that
didn't make it into the RN template-app commit. It looks like the
template app caught up with those when it bumped to 7.3, in
facebook/react-native@c180627ac.