msglist: Send message-now-unread events to the inside-webview code
This fixes a bug: if you're looking at a message list in the
mobile app and then mark some of those messages as unread (from
another client), they wouldn't regain their unread-markers.
The "N unread messages" banner at the top *would* update, but
the individual messages would still appear to be read.

Effectively this was a case where the props change in a way that
should describe a change in what's shown; we're preventing React
from doing its job to update things, taking responsibility for
doing it ourselves through the inbound-events mechanism; and we
had a gap in this code for doing that.

The old logic was correct in the past, when it was impossible for
a read message to become unread.  So from another angle, the root
cause here is that this is one of the many places across the Zulip
project where that longstanding invariant had been embedded as an
assumption, and so needed revision when we started breaking that
former invariant.

Fixes: #5536