i18n: Update for new Transifex client
Transifex is turning off their old API, and in the process they're
abandoning their longtime CLI client.  In its place is a new one,
written in Go.  ("For speed", though hilariously it's actually
a bit slower.)  Discussion here:
  https://chat.zulip.org/#narrow/stream/58-translation/topic/Transifex.20client/near/1460621

Update maintainer-facing docs to reflect the new client, and our
scripts and config to be compatible with it.  The config changes
came from the handy `tx migrate` provided in the new client.

Maintainers should read the new setup instructions and migrate
accordingly:

 * Install the new `tx` client.

 * Update the format of the `.transifexrc` file.  (The new `token`
   is the old `password`.)

 * Move the `.transifexrc` file to `~/`, if it was in the worktree.