android: Override MainActivity.invokeDefaultOnBackPressed, following Expo
This belatedly follows a change in Expo's template app,
templates/expo-template-bare-minimum. The change was made in
expo/expo@2a0079132, and we missed it in the Expo 44 upgrade (#5441)
because Expo's upgrade guide forgot about it. See discussion of this
and other changes that were missed:
  https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/Expo.2044.20small.20things/near/1452850

This seems to fix a bug I've observed on the office Android device
(Samsung Galaxy S9 running Android 9), where if I exit the app via
the back button, then re-enter the app, the state gets reset such
that we register for a new event queue.