ios: Follow Expo 44 in calling a super method in AppDelegate for Linking
This belatedly follows a change in Expo's template app,
templates/expo-template-bare-minimum. The change was made in
expo/expo@d20594b55 and expo/expo@b59dbc4d8, and we missed it in the
Expo 44 upgrade (#5441) because Expo's upgrade guide forgot about
it. See discussion of this and other changes that were missed:
  https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/Expo.2044.20small.20things/near/1452850

From reading the Expo commits and their associated PRs, I don't feel
like I understand the purpose of this change. But Greg says it
doesn't seem unreasonable:
  https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/Expo.2044.20small.20things/near/1458126
And nothing seemed broken when I tested the change on my iPhone 13
Pro running iOS 16.1. In particular, since we have a comment
mentioning social auth, I tested that I could still log in with
Google, and that worked.

Those Expo commits also touched a different method, meant to
implement an Apple feature called "Handoff":
  https://support.apple.com/en-us/HT209455
Here, we ignore the changes to that method; we haven't implemented
it, and while Handoff could sometimes be useful, it isn't a
priority. Also, Greg is skeptical about Expo's template-app changes
to that method.