queueMarkAsRead: Stop using server's misleading "messages" list
It turns out that the actual behavior of this response property is
that it just echoes the same message IDs we listed in the request.
So we might as well use the list we sent, instead.

Moreover, if this did tell us only those messages that were actually
affected, then among the messages left out would be messages that a
retry would have no better shot at affecting: messages that are
already read, or have been deleted.

(In fact on current server behavior those are the *only* messages
that would be left out -- there is no server limit on the size of
this request, so if it succeeds at all then it will have applied
to all the given messages that such a request could apply to.)

So by retrying those, we'd end up retrying in a loop forever.
That means there's no plausible future API change where consulting
this "messages" response property as we do here would be better
than ignoring it; and in fact our plan:
  https://chat.zulip.org/#narrow/stream/378-api-design/topic/mark-as-unread.20request/near/1463920
is to remove it in the future.  Start ignoring it now.