msglist [nfc]: Make more and narrower useSelector calls
This way useSelector is checking for equality on the individual props
we actually get from the state, rather than on the object we build
from all of them.  That gives it a better shot at avoiding re-renders
when nothing relevant has changed, though it's still defeated at that
by getBackgroundData.

See discussion:
  https://github.com/zulip/zulip-mobile/pull/5545#discussion_r1020523764