i18n: Sync recently-added message strings across languages.
This cherry-pick is a bit imprecise, because it adds strings that
exist in main but don't actually exist at the point we've
cherry-picked this onto.

But it's only adding strings, not removing any.  And having extra
strings is harmless.  So let them be; extracting only the relevant
strings would be more of a pain, and isn't needed.