unread: Fix VirtualizedList performance on un/collapse of stream
With the initial version of the collapsible-streams feature as added
in the previous commit, the UI can be extremely slow upon collapsing
a stream when the user has many unread topics, showing a glitchy
partially-blank state for 10 seconds or more.  Discussion starting at:
  https://github.com/zulip/zulip-mobile/pull/5467#issuecomment-1333137085

The unreads screen uses RN's SectionList, which is a thin wrapper
upon VirtualizedSectionList, which is a more complex wrapper upon
VirtualizedList, which does most of the work.  VirtualizedList, in
turn, makes an opaque -- and probably not entirely coherent -- set of
assumptions in its algorithms.

Upon some debugging, it seems like one of those assumptions is that
there aren't going to be lots and lots of list items with an extremely
small height like zero -- which is exactly what we do by keeping the
topics of a collapsed stream in the list but rendering them as null.

Instead, when a stream is collapsed, drop its topics from the list.
Empirically this gets us back into the range of acceptable performance:
still not *good*, when there are tons of unreads, but not worse than
the experience of trying to scroll past such a stream, as one would do
without the collapsible-streams feature.