api types: Shorten name of "cooked" response to ServerSettings
The type for the raw, unprocessed payload from the server is now
just named ApiResponseServerSettings, from
ServerApiResponseServerSettings.

The type for the "cooked" payload, with our validation on
`realm_name`, now has the shorter name ServerSettings, from
ApiResponseServerSettings.

The validation is done in a new helper function, `transform`. We'll
add more stuff there soon, and change the ServerSettings type to be
a more "cooked" form of the data.