deps: Drop `expo-cli`, the legacy Expo CLI
We don't use the `expo` CLI in normal development; only when
upgrading Expo itself.  Discussion:
  https://github.com/zulip/zulip-mobile/pull/5602#issuecomment-1347644361

This `expo-cli` package is apparently in maintenance mode.
It doesn't and won't support Node 18, the current LTS:
  https://github.com/expo/expo-cli/issues/4583
  https://github.com/expo/expo-cli/pull/4281

That effectively gives it an EOL of 2023-09, the EOL date of Node 16,
which is less than a year away.  It also means an error when just
trying to run `yarn install` with a newer Node version.  So it'd be
good to get it out of our dependencies.

There's a successor CLI now:
  https://blog.expo.dev/the-new-expo-cli-f4250d8e3421
That one is shipped with Expo itself, starting with Expo 46.
We're on Expo 45 at the moment, so we still have one more upgrade
before then; when we go to do that upgrade, we might need some
ad-hocery to get the CLI (either the old one or the new one,
which IIUC was in beta in the Expo 45 period.)  Until then,
just drop it.

As a bonus, dropping this sheds about 90MB from our node_modules tree.
(At least for now; maybe Expo 46 will bring that all back.)