android [nfc]: Set android:exported explicitly on DevSettingsActivity
This follows a change in the RN template app leading to RN v0.68:
facebook/react-native@00ac03435 Bump OSS Android build to SDK 31

Because this activity doesn't have any intent filters, this attribute
already has "false" as a default value, and there's no need to set it
explicitly:
  https://developer.android.com/guide/topics/manifest/activity-element#exported

SDK 31, aka Android 12, started requiring an explicit setting for this
attribute when there *is* an intent filter (and the app targets SDK 31+):
  https://developer.android.com/about/versions/12/behavior-changes-12#exported
(I think the default in that case had been "true", which is unsafe
when not intended.)  So in fa6fbbf7a we added it to our two other
activities, which do have intent filters.

It isn't really needed here; not only is the default what we want, but
if this activity somehow ever grew an intent filter then there'd be a
lint clearly pointing out that we needed this attribute.  But the
explicitness doesn't hurt anything, and keeps us aligned with the
template app; add it.