narrows tests: Fix type error missed by Flow
`initialState` wasn't a valid NarrowsState, which is

  Immutable.Map<string, $ReadOnlyArray<number>>

For what that type-checking problem is and how we plan to avoid it
in the future, see the next commit.

While we're here, make the test data a bit more interesting than the
intended value, by having the action act on a narrow that's
different from the one already in the state.