DebugScreen: Remove, as it's empty, and remove Redux plumbing for it
We're about to add a per-account troubleshooting screen for
notifications, and link to it from a better location: the
notification settings screen. Best to remove this empty screen so
that people don't get distracted by it when they want to
troubleshoot notifications.

This does remove a property from `SessionState`, but we don't need a
migration because 'session' is in `discardKeys`.