ProfileScreen: Use `NavRow`s and a `TextRow` instead of `ZulipButton`s
We've been using a `NavRow` and a `SwitchRow` for the first two
interactive items ("Set your status" and "Invisible mode"), and I
think the screen looks more coherent if we use `NavRow`s and a
`TextRow` for these other interactive items too. For one thing, now
all the text on the screen is the same color.