LegalScreen: Turn into a global, all-accounts screen
The user should be equally aware of the terms for all realms the app
is interacting with, and that includes realms other than the active
account's realm. (There would be even more interaction in a future
with better multi-account support.)

Rather than make the user traverse all realms with the "active
account" pointer to find all the relevant terms, it seems nicer to
show them together on this one page.