reactNativeUtils: Add 'unknown' branch to AppStateValues
As of writing, we've seen this reported to Sentry from two installs
of the app on iOS 16.2.

useAppState's callers don't try to take meaning from the returned
value, they just use it in a useEffect dependency array. So they
don't have to grow special logic to handle the new branch.