ios: Cherry-pick change to enforce iOS 14+ requirement
This cherry-picks e93d47f3d, which landed in `main` as a fix for
issue #5672 ("App Store thinks we support iOS 12 when we don't").

This is planned for a v27.203 release, to follow the recent v27.202
release, which fixed the iOS 12 bug #5670 after we realized our
mistake in letting #5672 happen.