ios: Make target Minimum Deployment inherit from project Deployment Target
The project's "iOS Deployment Target" can be seen in Xcode at:
  PROJECT > ZulipMobile > Info > Deployment Target > iOS Deployment Target
(This is already mentioned at the top of our Podfile.)

The *target*'s "Minimum Deployment" for iOS can be seen in Xcode at:
  TARGETS > ZulipMobile > General > Minimum Deployments > iOS

In 783e56b23, when we meant to bump the minimum supported iOS
version to 14, we neglected to touch the latter. This probably
caused #5672, "App Store thinks we support iOS 12 when we don't".

(I went and checked that latter value after seeing
https://developer.apple.com/forums/thread/688301, which I came
across while investigating investigating #5672.)

So we want to make that value 14. We could just change the Xcode
config lines to say 14 instead of 12, but we found we could do
better: have this value (the target's value) inherit from the
project's value. That's possible because, handily, they're both
"build settings", and they share the same identifier,
IPHONEOS_DEPLOYMENT_TARGET. Make that inheritance happen by deleting
the lines.

For how to confirm that this inheritance works as expected, see
Apple's helpful guide at
  https://developer.apple.com/documentation/xcode/configuring-the-build-settings-of-a-target#Evaluate-how-your-project-inherits-settings
It shows how to see a table of the target's build settings, showing
a "Resolved" column at left, and columns to the right, representing
nodes in an inheritance chain. After this change, 14 is in the
"Resolved" column, it's (newly) absent in the target's column, it's
absent in a "Config.File" column, and it's present as 14 in the
project's column.

Fixes: #5672